requirements.txt
requests
aiogram
aiofiles
python-dotenv